[router] move stages to different folder for grpc pipelines
